backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-socketio>=5.10.0
python-multipart>=0.0.6
opencv-python>=4.8.0
numpy>=1.26.0
pytesseract>=0.3.10
Pillow>=10.1.0
sqlalchemy>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.0
websockets>=12.0
scikit-image>=0.21.0
matplotlib>=3.7.0
seaborn>=0.12.0 